Chore : Store 디렉토리 생성
LoginStore와 SocketStore를 /Store 내부로 이동했습니다.
